Add an example of Phoenix LiveView with a single file
The JavaScript is served from a CDN for convenience, so the versions of
Phoenix and LiveView have been locked to match this.

The ErrorView is implemented in this example to prevent the ErrorView
missing errors being logged.

The way the layout is implemented is probably not best practice! It
makes it simple to code up an example if there are any JS changes
though and minimises the number of modules required.